Zmiany modelu bazy danych

+++ Domain/Models/Znizka.cs
@@ -15,7 +15,7 @@ namespace Domain.Models
 
         public int IdZnizka { get; set; }
         public string NazwaZnizki { get; set; }
-        public float ProcentZnizki { get; set; }
+        public decimal ProcentZnizki { get; set; }
         public DateTime? DoKiedy { get; set; }
 
         public virtual ICollection<KlientZnizka> KlientZnizkas { get; set; }
